+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqDeleteTopic.java
@@ -6,7 +6,7 @@ package com.tosslab.jandi.app.network.models;
 public class ReqDeleteTopic {
     public int teamId;
 
-    public ReqDeleteTopic(int teamId) {
+    public ReqDeleteTopic(long teamId) {
         this.teamId = teamId;
     }
 }
