+++ app/src/main/java/com/tosslab/jandi/app/ui/entities/chats/model/ChatChooseModel.java
@@ -134,7 +134,7 @@ public class ChatChooseModel {
 
     }
 
-    public int getTeamId() {
+    public long getTeamId() {
         return AccountRepository.getRepository().getSelectedTeamId();
     }
